chore(app): move toast to hook to avoid double render
